first working iteration of users api

+++ src/main/java/G2/SafeSpace/repository/UserRepository.java
@@ -3,4 +3,4 @@ package G2.SafeSpace.repository;
 import G2.SafeSpace.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {}
+public interface UserRepository extends JpaRepository<User, Integer> {}
